Generate correct server path in bindings
Problem:
If an interface API is used both as internal and external API then
server path in the bindings does not get generated correctly. The server
path should be in the format exe.compoonent.interface. Only interface
name gets printed in the server side in the bindings.

Resolution:
If an interface is of type internal and is also found listed in extern
server interfaces then generate full server path for the internal i.e.
in the format of exe.component.interface.

Resolves: LE-16223
Change-Id: I3c5ee2c8ec7d67c81e598a60fb75978daa4085dc