Add function to get JSON value start location
In some cases Octave needs to extract a sub-component from a JSON document.  In
order to do this, Octave needs to retrieve the offset and length of any item in
a JSON file or buffer.  To assist with this, provide a function
le_json_GetValueOffset() which gets the start and length of the most recently
parsed value.

Resolves: LE-16295
Change-Id: I8de79b2a06ffee0251f10ce63e17aa033d0e2c42