Fix race condition in FTP client library.
Prior to this patch, when we initiate a passive FTP data retrieve, the
FTP client would first issue a RETR (retrieve) command and then connect
to the FTP server's data port. When executing the happy path, where the
file path parameter RETR exists on the remote, the implementation is OK.
Problems arise when the file path is non-existent; the server may close
the data port upon receipt of the invalid path BEFORE the client has a
change to attempt a TCP connection to it. This race condition manifests
itself as unpredictable error codes being received on the client side:

 - 90% of the time the FTP server tested against
(ftp5.sierrawireless.com) would close the data socket before the client
connect()'s to it resulting in a TCP connection reset error
 - 10% of the time FTP client connect()'s to the server's data port
before the server has a change to reset the connection and as a result
the client gets back the FTP 550 (file not found) error code - this is
the expected behaviour

This patch reverses the order of the operations such that the FTP client
always first connects to the FTP server's data port BEFORE it issues the
FTP retrieve request.

This patch also adds several log messages to netSocket_Connect in order
to more easily diagnose connection failures in the future.

Resolves: ATSWI-235
Change-Id: I2727d37648e096d5f4cf70a9e0ff27298cd7dc03