Load framework config before processing sdef
In LE-16328 the location where interface search paths were added was
moved later in the sdef processing path to allow platform-specific
interfaces to be overridden in .sdef files.  However this had the
side-effect of including the framework configuration variables later in
processing as well, breaking .sdef files which depended on conditionals
on these configuration.

This change splits where interface search paths are added from where the
framework configuration is read and moves the framework configuration
back to its original location.

Resolves: LE-16371
Change-Id: Iff0f47f44534aef2b7850735f35ff08a224e99e5