Check if device is in data mode at the right place
Previously le_port_SetDataMode() does not return LE_DUPLICATE because it
did not check whether the device is already in the Data Mode at the
right place: it checked if we could disable fdMonitor for the device,
which should never fail.

Since we store the Data Mode file descriptor in the context variable, by
checking if the device has a valid Data Mode file descriptor, we can
accurately determine if the device is under the Data Mode.

Resolves: LE-13923
Change-Id: Id17b1d600bc63938a565b58373f5b1d2795691d5