Add a check for valid fdMonitor reference before enabling it
In same cases such as MQTT broker turns off the service before
client ends first, a POLLRDHUP will trigger the fdMoniter in
le_socketLib to delete and reset to NULL, but if client tries to
read again, le_socketLib will try to enable the fdMonitor again
which is no longer valid. Adding a check before enabling it fixes
an assertion exit.

This patch also

1. Optimizes the event handler of le_mqttClientLib to not
   continuously read from network when a POLLRDHUP is received.

2. Fixes an issue when TCP connection is successful but MQTT
   connection is not (e.g., connecting to a TCP only server), some
   clean up was not done. So, adding le_mqttClient_StopSession()
   when MQTTConnect() failed.

Resolves: LE-16434
Change-Id: I1eddf4868cd1c13607844b321da98f4ecbb21d58