Return ERROR when connecting MQTT broker with wrong password
When connecting to broker using wrong passsword, the embedded paho
lib (specifically the function MQTTConnect()) will return result
code 5, which is decoded from a socket packet when expecting for
CONNACK and not handled by le_mqttClientLib. This leads to
incorrectly converting the result code to a LE_OK to KMQTT and
hence no ERROR is output. Thus, enhancing the result code
conversion logic fixes this issue.

Resolves: LE-16322
Change-Id: I931ff95a48edf8cfa5f6b8a0dfa6f56a2c97d272