Correct set-after-use logic for buildVars
mkTools should give an error if a variable is set after it has been
used.  The original code had a bug where it would not properly handle
some cases where a variable was declared in a different file from
where it was used, and would instead consider it "lookahead" use,
and ignore it.  Since there is no lookahead across file boundaries,
always give an error if the files differ.

Also give a fatal error if errors cannot be recovered by jumping ahead
tokens.  This avoids an error cascade that can occur in some cases.

Resolves: LE-16701
Change-Id: I4597d609ecf07a8dffb8234a95622c1d75d979de