Create "fast reference" API
The current safe reference API creates a data structure which stores
the safe reference <-> pointer mapping.  However this is not always
necessary -- sometimes the objects themselves are stored elsewhere
and all that's needed is a method to check if an object is the object
refered to by the reference.

So create a le_ref_CreateFastRef() API which will create a reference
which is simply a kind of hash of the pointer.  These can be compared
to each other, but the object pointer cannot be looked up based
on a reference.

Resolves: LE-16717
Change-Id: I93c59194fd6d08ce5351bbf9b16cc02c1658fdec