Rework dataConnectionService unit test to single threaded
The dataConnectionService itself is a single-threaded application, but
the unit test was multithreaded.  This put some additional synchronization
demands on DCS soley for the unit test.  Rework the unit test to
continue to simulate multiple clients, but within a single thread.  This
makes it more similar to the actual application.

Resolves: LE-16717
Change-Id: Ie5d7dde0302e94fea2c66a720579a89261384ed3