Label Legato files for selinux
Add a '-c' option to pass a file containing a set of selinux labels
to mklegatoimg. This change requires support for -context-file
argument in mksquashfs, currently provided only in EM92xx targets.
Support in mkfs.ubifs is provided with --selinux flag.

Refer to setfiles man page for required syntax of label file.

Resolves: QTIX65-491
Change-Id: I5bc286466d4f08acaeba57964a32b772f6d4884d
Signed-off-by: Zoran Markovic <zmarkovic@sierrawireless.com>