Update cgroup management to co-exist with systemd
systemd uses cgroups to hierarchically group and label processes. So when
systemd is enabled, the cgroup hierarchy and all enabled controllers are
automatically mounted at startup by sytemd. In such case, Legato should
not remount the cgroup hierarchy, since that will result in losing
important configurations by systemd including the access to
/sys/fs/cgroup/systemd directory. This commit updates cgrp_Init() to
handle this case.

Also, with systemd using cgroups, the code to retrieve appname from
/proc/<pid>/cgroup file can no more rely on just the first line. It
needs to search for the entries for controllers that Legato configures.

Resolves: LE-16709
Change-Id: I512e42978950ba24f87ae9505d6114e25d09ebf0
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>